create-diff-object: compare section name with kpatch_section_function_name
Profile-Guided Optimization (PGO) uses profiling data to help the compiler
to evaluate the properties of a function, and thus adds different prefixes
to the section/function. For example, with -ffunction-sections, the
compiler will prefix some sectiones with .text.unlikely. However, if a
function changes from the version in the profiling data, the compiler may
ignore the profiling data. This often happens to the patched function
when kpatch-build builds the patch. As a result, the original function
and the patch function may have different prefix, i.e., one of them has
.text, the other has .text.unlikely.

To correlate these functions properly, use kpatch_section_function_name()
in kpatch_correlate_sections() to find matching sections.

Signed-off-by: Song Liu <songliubraving@fb.com>